Bit of a hack; get the IP into my X buffer
This will take the last output line, and take the fourth entry from that
line, and will put that information into my X buffer's CLIPBOARD
selection.  I'm pretty sure this is what I use everywhere else so it
should save me a lot of time copy/pasting the results on the linux side
of things.