Script to fix the JSAT-admt1 vpn alert
JSAT-admt1 FIT VPN - Peer IP 163.118.252.61
JSAT-admt1 10.2.199.21
https://nagios/cgi-bin/extinfo.cgi?type=1&host=jsat-admt1
https://nagios/cgi-bin/status.cgi?navbarsearch=1&host=10.2.199.21
https://nagios/cgi-bin/extinfo.cgi?type=2&host=jsat-admt1&service=FIT+VPN+-+Peer+IP+163.118.252.61

This is an expect script to accomplish the effects of the procedure
below.  This is the procedure from the extended notes in nagios.
Requires the neo script be installed/working.

The peer device has an issue where it will randomly drop this tunnel,
and silently discard any traffic destined for it until the JSAT firewall
rebuilds the tunnel, or the SPI expires (8 hours). To force this sooner,
we can just run "clear isa sa 163.118.252.61" on the firewall, then ping
172.16.22.129 from admt1. If it doesn't respond after a few seconds,
send an outage notification to monitoralerts@jsatech.com, and ACK it. No
further action is required from us unless JSA tells us otherwise.