Cannot use this kind of regex for a catchall
This keeps matching on one of the password entry prompts, either
the second one where it asks the new password or the third one
where it asks for that same new password again (I can't remember
which).  Anyway, there might be some kind of way to do a regex
that has negative lookarounds but I haven't figured that out yet.
So this should not match anything in the meantime and will serve
as a placeholder, and a reminder, to get back to this later.